refactor(secrets): enforce name validation in helper function
