+++ benefits/secrets.py
@@ -1,11 +1,13 @@
 import logging
 import os
+import re
 import sys
 
 from azure.core.exceptions import ClientAuthenticationError
 from azure.identity import DefaultAzureCredential
 from azure.keyvault.secrets import SecretClient
 from django.conf import settings
+from django.core.validators import RegexValidator
 
 logger = logging.getLogger(__name__)
 
@@ -13,11 +15,38 @@ logger = logging.getLogger(__name__)
 KEY_VAULT_URL = "https://kv-cdt-pub-calitp-{env}-001.vault.azure.net/"
 
 
+class SecretNameValidator(RegexValidator):
+    """RegexValidator that validates a secret name.
+
+    Azure KeyVault currently enforces the following rules:
+
+    * The value must be between 1 and 127 characters long.
+    * Secret names can only contain alphanumeric characters and dashes.
+
+    Read more about Azure KeyVault naming rules:
+    https://learn.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftkeyvault
+
+    Read more about Django validators:
+    https://docs.djangoproject.com/en/5.0/ref/validators/#module-django.core.validators
+    """
+
+    def __init__(self, *args, **kwargs):
+        kwargs["regex"] = re.compile(r"^[-a-zA-Z0-9]{1,127}$", re.ASCII)
+        kwargs["message"] = (
+            "Enter a valid secret name of between 1-127 alphanumeric ASCII characters and the hyphen character only."
+        )
+        super().__init__(*args, **kwargs)
+
+
+NAME_VALIDATOR = SecretNameValidator()
+
+
 def get_secret_by_name(secret_name, client=None):
     """Read a value from the secret store, currently Azure KeyVault.
 
     When `settings.RUNTIME_ENVIRONMENT() == "local"`, reads from the environment instead.
     """
+    NAME_VALIDATOR(secret_name)
 
     runtime_env = settings.RUNTIME_ENVIRONMENT()
