+++ tests/pytest/core/test_models.py
@@ -1,9 +1,9 @@
 from django.conf import settings
-from django.core.exceptions import ValidationError
 
 import pytest
 
-from benefits.core.models import SecretNameValidator, SecretValueField, EligibilityType, EligibilityVerifier, TransitAgency
+from benefits.core.models import SecretValueField, EligibilityType, EligibilityVerifier, TransitAgency
+import benefits.secrets
 
 
 @pytest.fixture
@@ -12,45 +12,10 @@ def mock_requests_get_pem_data(mocker):
     return mocker.patch("benefits.core.models.requests.get", return_value=mocker.Mock(text="PEM text"))
 
 
-@pytest.mark.parametrize(
-    "secret_name",
-    [
-        "a",
-        "1",
-        "one",
-        "one-two-three",
-        "1-2-3",
-        "this-is-a-really-long-secret-name-in-fact-it-is-the-absolute-maximum-length-of-127-characters-to-be-exact-and-now-it-has-enough",  # noqa: E501
-    ],
-)
-def test_SecretNameValidator_valid(secret_name):
-    validator = SecretNameValidator()
-
-    # a successful validation does not raise an Exception and returns None
-    assert validator(secret_name) is None
-
-
-@pytest.mark.parametrize(
-    "secret_name",
-    [
-        "",
-        "!",
-        "underscores_not_allowed",
-        "this-is-a-really-long-secret-name-in-fact-it-much-much-longer-than-the-absolute-maximum-length-of-127-characters-and-now-it-has-enough-to-be-too-long",  # noqa: E501
-    ],
-)
-def test_SecretNameValidator_invalid(secret_name):
-    validator = SecretNameValidator()
-
-    # an unsuccessful validation raises django.core.exceptions.ValidationError
-    with pytest.raises(ValidationError):
-        validator(secret_name)
-
-
 def test_SecretValueField_init():
     field = SecretValueField()
 
-    assert SecretValueField.NAME_VALIDATOR in field.validators
+    assert benefits.secrets.NAME_VALIDATOR in field.validators
     assert field.max_length == 127
     assert field.blank is False
     assert field.null is False
